requirements.txt
mysql-connector-python==8.0.30
pyTelegramBotAPI==4.7.0
python-dotenv==0.21.0
requests==2.28.1
